requirements.txt
discord.py==2.3.2
google-generativeai
python-dotenv
